renamed from cmd to dataset-filter
